Move parser to separate module
